refactor: move system stats from api to web data layer
Relocate the read path (getSystemStats/computeStats) out of the disabled
Hono /stats route into apps/web/src/lib/data/stats.ts as a cached server
data function with direct Postgres + Redis access. Keep the increment
counters in the API for release-check.ts.
