--- apps/api/src/routes/stats.ts
@@ -1,9 +1,0 @@
-import { Hono } from "hono";
-import { getSystemStats } from "../services/stats.service";
-
-const app = new Hono().basePath("/stats").get("/", async (c) => {
-  const stats = await getSystemStats();
-  return c.json(stats);
-});
-
-export default app;
